requirements.txt
# Core RAG (versions tương thích)
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-text-splitters==0.3.2
sentence-transformers==2.7.0

# Vector Database - Qdrant (Enterprise)
qdrant-client==1.12.1

# PDF Processing
pdfplumber==0.11.4

# MySQL Database
pymysql==1.1.0

# LLM
openai==1.30.1

# API Layer (Phase 2)
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.20

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# UI (Phase 3)
streamlit==1.41.1
plotly==5.24.1